Add stream resource cleanup
We defer the removal after a period to ensure that we've received all 
the final TCP close handshake packets for the session. At the point of 
closure, there should be no more data. If we don't do this, and we close 
on stream close events, we'll create dummy duplicate sessions, which we 
don't want.